refactor: Setting, FCM 도메인 수정
