+++ domain/src/main/kotlin/com/wespot/user/Setting.kt
@@ -1,7 +1,6 @@
 package com.wespot.user
 
 data class Setting(
-    val id: Long,
-    val isEnableNotification: Boolean,
+    val isEnableNotification: Boolean = false
 ) {
 }
