+++ infrastructure/mysql/src/main/kotlin/com/wespot/user/entity/SettingJpaEntity.kt
@@ -1,0 +1,12 @@
+package com.wespot.user.entity
+
+import jakarta.persistence.*
+import org.jetbrains.annotations.NotNull
+
+@Embeddable
+class SettingJpaEntity(
+
+    @field: NotNull
+    val isEnableNotification: Boolean
+
+)
